+++ src/textual/_doc.py
@@ -33,11 +33,13 @@ def format_svg(source, language, css_class, options, md, attrs, **kwargs) -> str
         try:
             rows = int(attrs.get("lines", 24))
             columns = int(attrs.get("columns", 80))
+            hover = attrs.get("hover", "")
             svg = take_svg_screenshot(
                 None,
                 path,
                 press,
-                title,
+                hover=hover,
+                title=title,
                 terminal_size=(columns, rows),
                 wait_for_animation=False,
             )
@@ -58,6 +60,7 @@ def take_svg_screenshot(
     app: App | None = None,
     app_path: str | None = None,
     press: Iterable[str] = (),
+    hover: str = "",
     title: str | None = None,
     terminal_size: tuple[int, int] = (80, 24),
     run_before: Callable[[Pilot], Awaitable[None] | None] | None = None,
@@ -69,6 +72,7 @@ def take_svg_screenshot(
         app: An app instance. Must be supplied if app_path is not.
         app_path: A path to an app. Must be supplied if app is not.
         press: Key presses to run before taking screenshot. "_" is a short pause.
+        hover: Hover over the given widget.
         title: The terminal title in the output image.
         terminal_size: A pair of integers (rows, columns), representing terminal size.
         run_before: An arbitrary callable that runs arbitrary code before taking the
@@ -97,7 +101,7 @@ def take_svg_screenshot(
             assert path is not None
             with open(path, "rb") as source_file:
                 hash.update(source_file.read())
-        hash.update(f"{press}-{title}-{terminal_size}".encode("utf-8"))
+        hash.update(f"{press}-{hover}-{title}-{terminal_size}".encode("utf-8"))
         cache_key = f"{hash.hexdigest()}.svg"
         return cache_key
 
@@ -115,7 +119,11 @@ def take_svg_screenshot(
             result = run_before(pilot)
             if inspect.isawaitable(result):
                 await result
+        await pilot.pause()
         await pilot.press(*press)
+        if hover:
+            await pilot.hover(hover)
+            await pilot.pause(0.5)
         if wait_for_animation:
             await pilot.wait_for_scheduled_animations()
             await pilot.pause()